raspberrypi/requirements.txt
requests
pyserial
datetime
argparse
termcolor